+++ gemma-rest/src/main/java/ubic/gemma/rest/util/args/GeneArgService.java
@@ -6,42 +6,106 @@ import ubic.gemma.core.association.phenotype.PhenotypeAssociationManagerService;
 import ubic.gemma.core.genome.gene.service.GeneService;
 import ubic.gemma.core.ontology.providers.GeneOntologyService;
 import ubic.gemma.core.search.SearchException;
+import ubic.gemma.model.expression.designElement.CompositeSequenceValueObject;
 import ubic.gemma.model.genome.Gene;
 import ubic.gemma.model.genome.GeneOntologyTermValueObject;
 import ubic.gemma.model.genome.PhysicalLocationValueObject;
 import ubic.gemma.model.genome.Taxon;
 import ubic.gemma.model.genome.gene.GeneValueObject;
 import ubic.gemma.model.genome.gene.phenotype.valueObject.GeneEvidenceValueObject;
+import ubic.gemma.persistence.service.expression.designElement.CompositeSequenceService;
+import ubic.gemma.persistence.util.Filter;
+import ubic.gemma.persistence.util.Filters;
+import ubic.gemma.persistence.util.Slice;
+import ubic.gemma.persistence.util.Sort;
 
+import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
+import javax.ws.rs.BadRequestException;
+import javax.ws.rs.NotFoundException;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
-import java.util.stream.Collectors;
 
 @Service
 public class GeneArgService extends AbstractEntityArgService<Gene, GeneService> {
 
     private final PhenotypeAssociationManagerService phenotypeAssociationManagerService;
     private final GeneOntologyService geneOntologyService;
-
+    private final CompositeSequenceService compositeSequenceService;
 
     @Autowired
-    public GeneArgService( GeneService service, PhenotypeAssociationManagerService phenotypeAssociationManagerService, GeneOntologyService geneOntologyService ) {
+    public GeneArgService( GeneService service, PhenotypeAssociationManagerService phenotypeAssociationManagerService, GeneOntologyService geneOntologyService, CompositeSequenceService compositeSequenceService ) {
         super( service );
         this.phenotypeAssociationManagerService = phenotypeAssociationManagerService;
         this.geneOntologyService = geneOntologyService;
-    }
-
-    public Gene getEntityWithTaxon( GeneArg<?> arg, Taxon taxon ) {
-        return checkEntity( arg, arg.getEntityWithTaxon( service, taxon ) );
+        this.compositeSequenceService = compositeSequenceService;
     }
 
     /**
-     * @return all genes that match the value of the GeneArg.
+     * {@inheritDoc}
+     * @throws BadRequestException if more than one gene match the supplied gene argument
      */
-    public List<GeneValueObject> getValueObjects( GeneArg<?> arg ) {
-        return service.loadValueObjects( getEntities( arg ) );
+    @Nonnull
+    @Override
+    public Gene getEntity( AbstractEntityArg<?, Gene, GeneService> entityArg ) throws NotFoundException, BadRequestException {
+        List<Gene> matchedGenes = getEntities( entityArg );
+        if ( matchedGenes.isEmpty() ) {
+            return checkEntity( entityArg, null );
+        } else if ( matchedGenes.size() > 1 ) {
+            throw new BadRequestException( "Gene identifier " + entityArg + " matches more than one gene, supply a taxon to disambiguate." );
+        } else {
+            return matchedGenes.iterator().next();
+        }
+    }
+
+    /**
+     * Obtain a gene from a specific taxon.
+     * @throws BadRequestException if more than one gene match the supplied gene argumen in the given taxon
+     */
+    @Nonnull
+    public Gene getEntityWithTaxon( GeneArg<?> entityArg, Taxon taxon ) {
+        List<Gene> matchedGenes = getEntitiesWithTaxon( entityArg, taxon );
+        if ( matchedGenes.isEmpty() ) {
+            return checkEntity( entityArg, null );
+        } else if ( matchedGenes.size() > 1 ) {
+            throw new BadRequestException( "Gene identifier " + entityArg + " matches more than one gene in " + taxon.getCommonName() + ", use a different type of identifier such as an NCBI or Ensembl ID." );
+        } else {
+            return matchedGenes.iterator().next();
+        }
+    }
+
+    /**
+     * Obtain genes from a specific taxon.
+     */
+    public List<Gene> getEntitiesWithTaxon( GeneArg<?> genes, Taxon taxon ) {
+        return genes.getEntitiesWithTaxon( service, taxon );
+    }
+
+    /**
+     * Obtain genes from a specific taxon.
+     */
+    public List<Gene> getEntitiesWithTaxon( GeneArrayArg genes, Taxon taxon ) {
+        List<Gene> objects = new ArrayList<>( genes.getValue().size() );
+        for ( String s : genes.getValue() ) {
+            GeneArg<?> arg = ( GeneArg<?> ) entityArgValueOf( genes.getEntityArgClass(), s );
+            objects.add( getEntityWithTaxon( arg, taxon ) );
+        }
+        return objects;
+    }
+
+    public Slice<GeneValueObject> getGenes( int offset, int limit ) {
+        return service.loadValueObjects( null, service.getSort( "id", Sort.Direction.ASC ), offset, limit );
+    }
+
+    public Slice<GeneValueObject> getGenesInTaxon( Taxon taxon, int offset, int limit ) {
+        return service.loadValueObjects( Filters.by( service.getFilter( "taxon.id", Long.class, Filter.Operator.eq, taxon.getId() ) ), service.getSort( "id", Sort.Direction.ASC ), offset, limit );
+    }
+
+    /**
+     * @return a collection of Gene value objects..
+     */
+    public List<GeneValueObject> getGenesInTaxon( GeneArrayArg arg, Taxon taxon ) {
+        return service.loadValueObjects( getEntitiesWithTaxon( arg, taxon ) );
     }
 
     /**
@@ -64,17 +128,8 @@ public class GeneArgService extends AbstractEntityArgService<Gene, GeneService> 
      * @param taxon       the taxon to limit the search to. Can be null.
      * @return collection of physical location objects.
      */
-    public List<PhysicalLocationValueObject> getGeneLocation( GeneArg<?> arg, Taxon taxon ) {
+    public List<PhysicalLocationValueObject> getGeneLocationInTaxon( GeneArg<?> arg, Taxon taxon ) {
         return service.getPhysicalLocationsValueObjects( getEntityWithTaxon( arg, taxon ) );
-    }
-
-    /**
-     * @return a collection of Gene value objects..
-     */
-    public List<GeneValueObject> getGenesOnTaxon( GeneArg<?> arg, Taxon taxon ) {
-        return this.getValueObjects( arg ).stream()
-                .filter( vo -> Objects.equals( vo.getTaxonId(), taxon.getId() ) )
-                .collect( Collectors.toList() );
     }
 
     /**
@@ -92,11 +147,29 @@ public class GeneArgService extends AbstractEntityArgService<Gene, GeneService> 
 
     /**
      * Returns GO terms for the gene that this GeneArg represents.
-     *
-     * @return collection of physical location objects.
      */
-    public List<GeneOntologyTermValueObject> getGoTerms( GeneArg<?> arg ) {
-        Gene gene = this.getEntity( arg );
-        return geneOntologyService.getValueObjects( gene );
+    public List<GeneOntologyTermValueObject> getGeneGoTerms( GeneArg<?> arg ) {
+        return geneOntologyService.getValueObjects( this.getEntity( arg ) );
+    }
+
+    /**
+     * Obtain GO terms for the gene in the given taxon.
+     */
+    public List<GeneOntologyTermValueObject> getGeneGoTermsInTaxon( GeneArg<?> geneArg, Taxon taxon ) {
+        return geneOntologyService.getValueObjects( this.getEntityWithTaxon( geneArg, taxon ) );
+    }
+
+    /**
+     * Obtain probes for the gene across all platforms.
+     */
+    public Slice<CompositeSequenceValueObject> getGeneProbes( GeneArg<?> geneArg, int offset, int limit ) {
+        return compositeSequenceService.loadValueObjectsForGene( getEntity( geneArg ), offset, limit );
+    }
+
+    /**
+     * Obtain probes for the gene in the given taxon across all platforms.
+     */
+    public Slice<CompositeSequenceValueObject> getGeneProbesInTaxon( GeneArg<?> geneArg, Taxon taxon, int offset, int limit ) {
+        return compositeSequenceService.loadValueObjectsForGene( getEntityWithTaxon( geneArg, taxon ), offset, limit );
     }
 }